add delete user api
